article-search-api/requirements.txt
google-generativeai==0.8.4
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
uvicorn==0.32.1
fastapi==0.115.6
fastapi-cli==0.0.7